+++ packages/ui/VirtualList/tests/VirtualList-specs.js
@@ -320,5 +320,26 @@ describe('VirtualList', () => {
 				expect(actual).toBe(expected);
 			}
 		);
+
+		test(
+			'should render "data-webos-voice-disabled" to outermost node of VirtualList',
+			() => {
+				const subject = mount(
+					<VirtualList
+						cbScrollTo={getScrollTo}
+						clientSize={clientSize}
+						dataSize={dataSize}
+						itemRenderer={renderItem}
+						itemSize={30}
+						data-webos-voice-disabled
+					/>
+				);
+
+				const expected = true;
+				const actual = subject.find(`.${css.virtualList}`).prop('data-webos-voice-disabled');
+
+				expect(actual).toBe(expected);
+			}
+		);
 	});
 });